+++ README.md
@@ -70,6 +70,10 @@ Chat反诈模型全流程如图所示：
 
 ***2024年6月21日，完成 [InternLM2-chat-1_8B模型的微调](https://openxlab.org.cn/models/detail/Eric-sc/Fanzha1_8B/tree/main)***
 
+***2024年6月30日，更新数据集，并重新训练模型，重新评测，修复了模型在训练后失去推理能力的问题.***
+
+***2024年6月30日，更新页面展示效果***
+
 ## 数据集
 
 Chat-反诈模型 数据集采用各种网络诈骗案例和用户的聊天记录，共计 3000 余条,位于datasets文件夹下.数据集样例：
