+++ requirements.txt
@@ -1,5 +1,5 @@
 modelscope==1.9.5
-streamlit==1.32.0
+streamlit==1.36.0
 transformers==4.36.2
 sentencepiece==0.1.99
 accelerate==0.24.1
